+++ newtmgr/bll/bll_sesn_cfg.go
@@ -30,7 +30,7 @@ import (
 type BllSesnCfg struct {
 	MgmtProto    sesn.MgmtProto
 	AdvFilter    ble.AdvFilter
-	PreferredMtu int
+	PreferredMtu uint16
 	ConnTimeout  time.Duration
 	ConnTries    int
 }
